refactor: rename src/task/action-constants.ts to action-types.ts
